reorganize proxy code into shard.go
